Implemented BlockData updates
This version allows files from older versions of Minecraft to be updated
to the newest version.

This updates include changing the LRUCache to support strings, since the
updater now supports updating TO a version that is not the current
version. To do that, it must do it without using BlockData, since the
blockdata for the current version may not be compatible.

This version allows users to specify blockdata updates, allowing them to
use RollbackCore without issues on a technically unsupported version.

This version is tolerant of invalid blockdata, and ignores it properly.